refactor(cases): split case task list components
